Now using initial middleware
