+++ server/package.json
@@ -9,7 +9,9 @@
   "author": "Patrick Fisher <patrick@pwfisher.com>",
   "license": "MIT",
   "dependencies": {
-    "fastboot-app-server": "pwfisher/fastboot-app-server.git#user-middleware",
-    "fastboot-watch-notifier": "2.1.0"
+    "connect-datadog": "0.0.6",
+    "fastboot-app-server": "dollarshaveclub/fastboot-app-server.git#initial-middleware",
+    "fastboot-watch-notifier": "2.1.0",
+    "raven": "^0.12.1"
   }
 }
