+++ server/server.js
@@ -2,7 +2,6 @@
 
 const FastBootAppServer = require('fastboot-app-server');
 const FastBootWatchNotifier = require('fastboot-watch-notifier');
-const isFastBootRoute = require('./is-fastboot-route');
 
 const distPath = '/app/dist';
 
@@ -13,25 +12,13 @@ const notifier = new FastBootWatchNotifier({
   },
 });
 
-const userMiddleware = function (req, res, next) {
-
-  let doFastBoot = isFastBootRoute(req.path);
-  if (req.query.fastboot === 'on') doFastBoot = true;
-  if (req.query.fastboot === 'off') doFastBoot = false;
-
-  if (doFastBoot) {
-    next();
-  } else {
-    res.sendFile('index.html', { root: distPath });
-  }
-};
-
+const initialMiddleware = require('./middleware');
 const workerCount = process.env.WORKER_COUNT;
 
 const server = new FastBootAppServer({
   distPath,
   notifier,
-  userMiddleware,
+  initialMiddleware,
   workerCount,
 });
